refactor: extract tool-filters to separate module to prevent test hangs
Move TOOL_GROUPS and getDisabledToolNames to src/tool-filters.ts so
tests can import them without triggering the MCP server startup in
index.ts. This ensures npm test completes cleanly.

Agent-Logs-Url: https://github.com/dwin/fastmail-mcp/sessions/0562ca66-afa1-4cd4-bf4d-f73d54727662
